fix(diffusion): cast SC soft-embed to fp32 buffer dtype at store
The native DiffusionGemma sampler stores probs@embed_weight (bf16) into the
persistent self_conditioning_embeds buffer, which is allocated fp32. index_put_
requires matching source/destination dtypes, so serving crashed at the first
sample step (RuntimeError: Index put requires the source and destination dtypes
match, got Float for the destination and BFloat16 for the source) — crash-looping
EngineCore and blocking ALL diffusion serving, adapter or not. Cast the source to
the buffer dtype at store; the consumer already casts back on read.

+++ vllm/model_executor/models/diffusion_gemma.py
@@ -656,7 +656,11 @@ def _compiled_sample_step(
     if tp_size > 1:
         soft_embeds = torch.ops.vllm.all_reduce(soft_embeds, group_name=tp_group_name)
     soft_embeds = soft_embeds * normalizer
-    sc_embeds[decode_slots] = soft_embeds * sc_keep
+    # sc_embeds is a persistent fp32 buffer (see DiffusionGemmaRequestStates)
+    # while soft_embeds is bf16 (probs @ embed_weight); index_put_ requires the
+    # source dtype to match the destination, so cast to the buffer dtype. The
+    # consumer (_apply_self_conditioning) symmetrically casts back on read.
+    sc_embeds[decode_slots] = (soft_embeds * sc_keep).to(sc_embeds.dtype)
 
     # Overwrite canvas with argmax for newly converged denoise requests
     newly_converged = (converged & is_denoise).unsqueeze(1)
